Fixing admin panel

+++ admin/sql_query_del/subscriber_del.php
@@ -1,16 +1,17 @@
 <?php
 
 require_once $_SERVER["DOCUMENT_ROOT"] . "/connect__bd/connect__bd.php";
-
-$stmt = mysqli_prepare($connect, "DELETE FROM `subscribe_user` WHERE `subscribe_user_id` = ? LIMIT 1;");
-mysqli_stmt_bind_param($stmt, "i", $subscribe_user_id);
-if (mysqli_stmt_execute($stmt)) {
-    mysqli_stmt_close($stmt);
-    echo "<script>alert('Удалено успешно!'); location.href = '/admin/index.php';</script>";
-    exit;
-} else {
-    mysqli_stmt_close($stmt);
-    echo "<script>alert('Не удалось!'); location.href = '/admin/index.php';</script>";
-    exit;
+if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST["subscribe_user_id"])) {
+    $subscribe_user_id = (int) $_POST["subscribe_user_id"];
+    $stmt = mysqli_prepare($connect, "DELETE FROM `subscribe_user` WHERE `subscribe_user_id` = ? LIMIT 1;");
+    mysqli_stmt_bind_param($stmt, "i", $subscribe_user_id);
+    if (mysqli_stmt_execute($stmt)) {
+        mysqli_stmt_close($stmt);
+        echo "<script>alert('Удалено успешно!'); location.href = '/admin/index.php';</script>";
+        exit;
+    } else {
+        mysqli_stmt_close($stmt);
+        echo "<script>alert('Не удалось!'); location.href = '/admin/index.php';</script>";
+        exit;
+    }
 }
-
